refactor(wave0d.1 #252): move notify deciders to daemon
Move electron/notify/{notifyDecider,runStateTracker,titleStateClassifier,
badgeLabel,badgePixels,badgeStore}.ts and their tests into
packages/daemon/src/notify/. All six are pure (node builtins only; the
electron API imports live in sinks/ and bootstrap/, which remain in
electron/).

Tests:
* electron/notify/__tests__/*.test.ts → packages/daemon/src/notify/
  __tests__/*.spec.ts
* tests/badgeLabel.test.ts → packages/daemon/src/notify/__tests__/
  badgeLabel.root.spec.ts (kept distinct from same-named module spec —
  the two cover different angles of the same module)
* tests/badgePixels.test.ts → ditto badgePixels.root.spec.ts

Re-export shims at electron/notify/{the six paths} keep
electron/notify/sinks/* (Wave 0c #217 just merged) and the bootstrap
pipeline compiling. Shim deletion happens once notify/sinks/ either
moves to daemon or is replaced by Connect-RPC subscriptions from the
renderer (flagged for a notify-cleanup task).

[REFACTOR-ONLY] no behavior change.
